requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
boto3==1.40.65
botocore==1.40.65
click==8.3.0
colorama==0.4.6
fastapi==0.121.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
idna==3.11
jmespath==1.0.1
packaging==25.0
pydantic==2.12.3
pydantic_core==2.41.4
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.14.0
setuptools==65.5.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
